requirements.txt
#Backend
langchain
python-dotenv
fastapi
langchain-groq
uvicchain-core
langchain-community
lanorn
langgchain-classic
langchain-huggingface
sentence-transformers
langchain-text-splitters
sentence-transformers
langchain-chroma
huggingface-hub
chromadb
docx2txt
transformers
pymupdf
rank-bm25
firecrawl
langchain-google-genai
google-generativeai
llama-cloud>=1.0
llama-parse

#Frontend
langchain
langchain-ollama
langgraph
chainlit
langchain-core
aiohttp
greenlet
asyncpg
SQLAlchemy
uvicorn
fastapi
